requirements.txt
absl-py==1.3.0
accelerate==0.14.0
aiofiles==23.1.0
aiohttp==3.8.3
aiosignal==1.3.1
altair==4.2.0
antlr4-python3-runtime==4.9.3
anyio==3.6.1
argilla==1.6.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asn1crypto==0.24.0
asttokens==2.0.8
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
Babel==2.10.3
backcall==0.2.0
backoff==2.2.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
biopython==1.68
biotite==0.35.0
bleach==5.0.1
blinker==1.5
cachetools==5.2.0
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.1
coloredlogs==15.0.1
commonmark==0.9.1
contextlib2==21.6.0
contourpy==1.0.7
cpm-kernels==1.0.11
cryptography==40.0.1
cycler==0.11.0
dataclasses-json==0.5.7
debugpy==1.6.3
decorator==5.1.1
deepspeed==0.5.9
defusedxml==0.7.1
DendroPy==4.5.2
Deprecated==1.2.13
deprecation==2.1.0
DLLogger @ git+https://github.com/NVIDIA/dllogger.git@0540a43971f4a8a16693a9de9de73c1072020769
dm-tree==0.1.7
docker-pycreds==0.4.0
effdet==0.3.0
einops==0.6.0
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.1.1
fair-esm @ git+https://github.com/facebookresearch/esm.git@a944dc53a9774e5ebfab4be1a395c2c516325e37
faiss-cpu==1.7.2
fastapi==0.95.0
fastjsonschema==2.16.2
ffmpy==0.3.0
filelock==3.8.0
fire==0.4.0
flatbuffers==23.3.3
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2022.11.0
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.29
google-auth==2.14.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gptcache==0.1.10
gradio==3.26.0
gradio_client==0.1.2
greenlet==2.0.2
grpcio==1.51.1
h11==0.14.0
h5py==3.8.0
hjson==3.1.0
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.13.2
humanfriendly==10.0
icetk==0.0.5
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.10.0
iopath==0.1.10
ipykernel==6.16.0
ipyspeck==0.6.1
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.10
jsonschema==4.16.0
jupyter-core==4.11.1
jupyter-server==1.19.1
jupyter_client==7.3.5
jupyter_packaging==0.12.3
jupyterlab==3.4.8
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
jupyterlab_server==2.15.2
keras==2.11.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.4.4
langchain==0.0.139
layoutparser==0.3.4
lazy_loader==0.1rc2
libclang==15.0.6.1
lightning-utilities==0.3.0
lightseq==3.0.1
linkify-it-py==2.0.0
lxml==4.9.2
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.1
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
mistune==2.0.4
ml-collections==0.1.1
monotonic==1.6
mpmath==1.3.0
msg-parser==1.2.0
msgpack==1.0.4
multidict==6.0.2
mypy-extensions==1.0.0
nbclassic==0.4.5
nbclient==0.7.0
nbconvert==7.2.1
nbformat==5.6.1
nest-asyncio==1.5.6
networkx==2.8.8
ninja==1.11.1
nltk==3.8.1
notebook==6.4.12
notebook-shim==0.1.0
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
olefile==0.46
omegaconf==2.2.3
onnxruntime==1.14.1
openai==0.27.4
openapi-schema-pydantic==1.2.4
opencv-python==4.6.0.66
openfold @ git+https://github.com/aqlaboratory/openfold.git@4b41059694619831a7db195b7e0988fc4ff3a307
openpyxl==3.1.2
opt-einsum==3.3.0
orjson==3.8.7
packaging==21.3
pandas==1.5.2
pandas-flavor==0.3.0
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.9.0
pexpect==4.8.0
phylopandas==0.8.0
pickleshare==0.7.5
Pillow==9.2.0
pkgutil_resolve_name==1.3.10
portalocker==2.7.0
prometheus-client==0.14.1
promise==2.3
prompt-toolkit==3.0.31
protobuf==3.18.3
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
py3Dmol==1.8.1
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.6
pycparser==2.21
pycrypto==2.6.1
pydantic==1.10.6
pydeck==0.8.0
pydub==0.25.1
Pygments==2.13.0
PyGObject==3.26.1
Pympler==1.0.1
pypandoc==1.11
pyparsing==3.0.9
pyrsistent==0.18.1
pytesseract==0.3.10
python-apt==1.6.5+ubuntu0.7
python-dateutil==2.8.2
python-docx==0.8.11
python-magic==0.4.27
python-multipart==0.0.6
python-pptx==0.6.21
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
pyxdg==0.25
PyYAML==6.0
pyzmq==24.0.1
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==12.6.0
rsa==4.9
scikit-learn==1.2.2
scipy==1.9.3
seaborn==0.12.2
SecretStorage==2.3.1
semantic-version==2.10.0
semver==2.13.0
Send2Trash==1.8.0
sentence-transformers==2.2.2
sentencepiece==0.1.97
sentry-sdk==1.11.1
setproctitle==1.3.2
shortuuid==1.0.11
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.47
ssh-import-id==5.7
stack-data==0.5.1
starlette==0.26.1
stmol==0.0.9
streamlit==1.13.0
supervisor==4.2.4
sympy==1.11.1
tenacity==8.2.2
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.1.1
terminado==0.16.0
threadpoolctl==3.1.0
timm==0.6.13
tinycss2==1.1.1
tokenizers==0.13.2
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
toolz==0.12.0
torch==1.13.1
torch-cluster==1.6.1+pt113cu117
torch-geometric==2.2.0
torch-scatter==2.1.1+pt113cu117
torch-sparse==0.6.17+pt113cu117
torch-spline-conv==1.2.2+pt113cu117
torchaudio==0.13.1
torchvision==0.14.1
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
transformers==4.27.1
triton==1.0.0
tvm==0.12.dev382+g6fa88e38a
typing-inspect==0.8.0
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
uc-micro-py==1.0.1
unattended-upgrades==0.1
unstructured==0.5.12
unstructured-inference==0.3.2
urllib3==1.26.12
uvicorn==0.21.1
validators==0.20.0
Wand==0.6.11
wandb==0.13.4
watchdog==2.1.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.1
websockets==10.4
Werkzeug==2.2.2
widgetsnbextension==3.5.2
wrapt==1.14.1
xarray==2022.11.0
XlsxWriter==3.1.0
yarl==1.8.1
zipp==3.9.0
